moved get_url to utils
