+++ dnora/read/grid/grid_readers.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import xarray as xr
-from abc import ABC, abstractmethod
 import numpy as np
 from typing import TYPE_CHECKING
 import json
@@ -11,7 +10,7 @@ import re
 if TYPE_CHECKING:
     from dnora.grid import Grid, TriGrid
 from dnora.read.abstract_readers import DataReader
-from dnora.aux_funcs import get_url
+from dnora.utils.io import get_url
 from dnora import utils
 from dnora import msg
 from typing import Union
