+++ dnora/read/wind/wind_readers.py
@@ -1,7 +1,7 @@
 from dnora.read.abstract_readers import DataReader
 from dnora.type_manager.data_sources import DataSource
 from dnora.type_manager.dnora_types import DnoraDataType
-from dnora.aux_funcs import get_url
+from dnora.utils.io import get_url
 from dnora.grid import Grid
 from dnora import msg
 from dnora import utils
